Make interactions more generic
